Fix qcow2 disk images for data sizes greater than 4 gigs. We used to truncate the disk end by anding it with a 32 bit value. The 32 bit value was not sign extended, which causes the disk size to wrap at 4 gigabytes:         disk->end = (disk->end + disk->clustersz - 1) & ~(disk->clustersz - 1); This change converts the clustersz to an off_t in order to remove the class of errors by avoiding type conversions entirely. from ori@
OpenBSD 6.4 errata 008

+++ usr.sbin/vmd/vioqcow2.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: vioqcow2.c,v 1.8 2018/10/08 16:32:01 reyk Exp $	*/
+/*	$OpenBSD: vioqcow2.c,v 1.8.2.1 2018/11/29 17:25:31 bluhm Exp $	*/
 
 /*
  * Copyright (c) 2018 Ori Bernstein <ori@eigenstate.org>
@@ -78,15 +78,15 @@ struct qcdisk {
 	int       fd;
 	uint64_t *l1;
 	off_t     end;
-	uint32_t  clustersz;
+	off_t	  clustersz;
 	off_t	  disksz; /* In bytes */
-	uint32_t cryptmethod;
+	uint32_t  cryptmethod;
 
 	uint32_t l1sz;
 	off_t	 l1off;
 
 	off_t	 refoff;
-	uint32_t refsz;
+	off_t	 refsz;
 
 	uint32_t nsnap;
 	off_t	 snapoff;
@@ -176,7 +176,7 @@ qc2_open(struct qcdisk *disk, int *fds, size_t nfd)
 	struct qcheader header;
 	uint64_t backingoff;
 	uint32_t backingsz;
-	size_t i;
+	off_t i;
 	int version, fd;
 
 	pthread_rwlock_init(&disk->lock, NULL);
